@@ -1,5 +1,5 @@
 {
-  "name": "reactwmw",
+  "name": "wmr",
   "version": "1.3.1",
   "private": true,
   "dependencies": {
@@ -44,12 +44,13 @@
     "react-i18next": "^11.11.4",
     "react-konva": "^17.0.2-5",
     "react-redux": "^7.2.5",
-    "react-router": "^6.2.1",
+    "react-router": "^6.3.0",
     "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^6.2.1",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "^3.4.4",
     "reactstrap": "^8.4.1",
     "recharts": "^2.1.9",
+    "redux": "^4.2.0",
     "rimraf": "^2.6.2",
     "svg2pdf.js": "^2.2.0",
     "three": "^0.139.2",
@@ -72,7 +73,9 @@
     "nan": "^2.14.1",
     "postcss": "^8.3.11",
     "postcss-loader": "^4.3.0",
-    "typescript": "^3.7.5",
+    "ts-loader": "^9.3.0",
+    "typescript": "^4.6.4",
+    "webpack": "^4.42.0",
     "webpack-cli": "^4.9.1"
   },
   "eslintConfig": {
