@@ -1,84 +1,105 @@
 ﻿import * as React from "react";
-import { useSelector } from 'react-redux';
-import { useTranslation } from 'react-i18next';
+import { useSelector } from "react-redux";
+import { useTranslation } from "react-i18next";
 import { ComboBox } from "@progress/kendo-react-dropdowns";
 import { stringToTableCell, tableRowToString } from "../../utils";
 
 
+/*
+props {
+  displayName: string,
+  editorType: string,
+  externalChannelName: string,
+  formId: string,
+  formIdToLoad: any,
+  id: string,
+  selectionChanged: function,
+  value: any,
+}
+*/
+
+/*
+KendoReact ComboBox props {
+  name: string, // задает свойство имени входного элемента DOM
+  data: any[], // список возможных вариантов
+  value: any, // значение ComboBox; может быть примитивным или сложным (напр. объект)
+  suggest: boolean, // автозаполнение текста на основе первого элемента данных
+  allowCustom: boolean, // возможность установить значение не из предлагаемых
+  placeholder: string, // подсказка, которая отображается, когда ComboBox пуст
+  onOpen: function, // срабатывает, когда всплывающее окно ComboBox вот-вот откроется.
+  onChange: function, // срабатывает каждый раз, когда значение ComboBox собирается измениться
+}
+*/
+
+
 export default function TableRowComboEditor(props) {
-    const { t } = useTranslation();
-    const { id, formId, selectionChanged, externalChannelName } = props;
-    let values = [];
-    let valueToShow = undefined;
-    const value = useSelector((state) => state.formParams[formId].find((gp) => gp.id === id).value);
-    const nullDisplayValue = useSelector((state) => state.formParams[formId].find((gp) => gp.id === id).nullDisplayValue ?? t("editors.activeObjectNullDisplayName"));
-    const showNullValue = useSelector((state) => state.formParams[formId].find((gp) => gp.id === id).showNullValue);
-    const sessionManager = useSelector((state) => state.sessionManager);
+  const { t } = useTranslation();
+  const { id, formId, selectionChanged, externalChannelName } = props;
+  let values = [], valueToShow = undefined;
 
-    const setNewValue = React.useCallback((value, manual) => {
-            const newEvent = {};
-            newEvent.target = {};
-            newEvent.target.name = id;
-            newEvent.target.manual = manual;
-            newEvent.target.value = value;
-            selectionChanged(newEvent);
-        },
-        [id, selectionChanged],
-    );
+  const sessionManager = useSelector((state) => {
+    // console.log('props', props);
+    // console.log('state', state);
+    return state.sessionManager;
+  });
+  const formParameter = useSelector((state) => state.formParams[formId].find((gp) => gp.id === id));
+  const valuesToSelect = useSelector((state) => state.channelsData[externalChannelName]);
 
-    const valuesToSelect = useSelector((state) => state.channelsData[externalChannelName]);
+  const value = formParameter.value;
+  const nullDisplayValue = formParameter.nullDisplayValue ?? t("editors.activeObjectNullDisplayName");
+  const showNullValue = formParameter.showNullValue;
 
-    if (valuesToSelect && valuesToSelect.properties) {
-        const valuesFromJSON = valuesToSelect?.data?.Rows?.map((row) => tableRowToString(valuesToSelect, row));
+  if (valuesToSelect && valuesToSelect.properties) {
+    const valuesFromJSON = valuesToSelect?.data?.Rows?.map((row) => tableRowToString(valuesToSelect, row));
 
-        values = (valuesFromJSON && valuesFromJSON !== '') ? valuesFromJSON : [];
+    values = (valuesFromJSON && valuesFromJSON !== '') ? valuesFromJSON : [];
 
-        if (showNullValue) {
-            values.push({id: null, name: nullDisplayValue, value: null})
-        }
-
-        if (value) {
-            const dataId = stringToTableCell(value, 'LOOKUPCODE');
-            const calculatedValueToShow = values.find(o => String(o.id) === dataId);
-            valueToShow = calculatedValueToShow ? calculatedValueToShow : '';
-        }
-        else if (showNullValue) {
-            valueToShow = {id: value, name: nullDisplayValue, value: value};
-        }
-    }
-    else if (value) {
-        valueToShow = {
-            id: stringToTableCell(value, 'LOOKUPCODE'),
-            name: stringToTableCell(value, 'LOOKUPVALUE'),
-            value: value
-        };
-    }
-    else if (showNullValue) {
-        valueToShow = {
-            id: value,
-            name: nullDisplayValue,
-            value: value
-        };
+    if (showNullValue) {
+      values.push({id: null, name: nullDisplayValue, value: null})
     }
 
-    const onOpen = () => {
-        sessionManager.channelsManager.loadAllChannelData(externalChannelName, formId, false);
+    if (value) {
+      const dataId = stringToTableCell(value, 'LOOKUPCODE');
+      const calculatedValueToShow = values.find(o => String(o.id) === dataId);
+      valueToShow = calculatedValueToShow ? calculatedValueToShow : '';
+    } else if (showNullValue) {
+      valueToShow = {id: value, name: nullDisplayValue, value: value};
+    }
+  } else if (value) {
+    valueToShow = {
+      id: stringToTableCell(value, 'LOOKUPCODE'),
+      name: stringToTableCell(value, 'LOOKUPVALUE'),
+      value: value
     };
+  } else if (showNullValue) {
+    valueToShow = {id: value, name: nullDisplayValue, value: value};
+  }
 
-    return (
-        <ComboBox className='parametereditor'
-            onOpen={onOpen}
-            suggest={true}
-            name={id}
-            data={values}
-            value={valueToShow}
-            dataItemKey="id"
-            textField="name"
-            placeholder={nullDisplayValue}
-            onChange={(event) => {
-                valueToShow = event.target.value;
-                setNewValue(event.target.value?.value, true);
-            }}
-        />
-    );
+  const [readyValueToShow, setReadyValueToShow] = React.useState(valueToShow);
+
+  const setNewValue = React.useCallback((value, manual) => {
+    selectionChanged({target: {name: id, manual, value}});
+  }, [id, selectionChanged]);
+
+  // event: ComboBoxChangeEvent
+  const onChange = (event) => {
+    setReadyValueToShow(event.target.value);
+    setNewValue(event.target.value?.value, true);
+  }
+
+  const onOpen = () => {
+    sessionManager.channelsManager.loadAllChannelData(externalChannelName, formId, false).then();
+  };
+
+  if (id === 'wellCurrent') console.log('values', values);
+
+  return (
+    <ComboBox
+      className='parametereditor' dataItemKey="id" textField="name"
+      name={id} data={values}
+      value={readyValueToShow} placeholder={nullDisplayValue}
+      suggest={true} allowCustom={true}
+      onChange={onChange} onOpen={onOpen}
+    />
+  );
 }
