@@ -3,29 +3,51 @@ import { useSelector } from 'react-redux';
 import { StackLayout } from "@progress/kendo-react-layout";
 import BaseEditor from '../activeParametersEditors/BaseEditor';
 
+
+/*
+param {
+  id: string,
+  type: string,
+  value: any,
+  formId: string,
+  editorType: string,
+  editorDisplayOrder: number, // в каком порядке отображать параметры
+  externalChannelName: any,
+  displayName: string,
+  dependsOn: string[], // id других параметров
+  canBeNull: boolean,
+  showNullValue: boolean,
+  nullDisplayValue: any,
+}
+*/
+
+/** Функция сортировки параметров. */
+const sortParams = (a, b) => a.editorDisplayOrder - b.editorDisplayOrder;
+
+
+/** Компонент списка параметров. */
 export default function ParametersList(props) {
-    const paramsManager = useSelector((state) => state.sessionManager.paramsManager);
+    const updateParamValue = useSelector((state) => state.sessionManager.paramsManager.updateParamValue);
     const { parametersJSON } = props;
 
     const updateEditedJSON = (action, formId) => {
         const target = action.target;
-        const newValue = action.value ?? target.value;
-        paramsManager.updateParamValue(formId, target.name, newValue, target.manual ?? true);
+        updateParamValue(formId, target.name, action.value ?? target.value, target.manual ?? true);
     };
 
     return (
         <StackLayout orientation="vertical">
-            {parametersJSON.filter(parameterJSON => parameterJSON.editorType).sort((a, b) => a.editorDisplayOrder - b.editorDisplayOrder).map(parameterJSON =>
+            {parametersJSON.filter(param => param.editorType).sort(sortParams).map(param =>
                 <BaseEditor
-                    editorType={parameterJSON.editorType}
-                    key={parameterJSON.id}
-                    id={parameterJSON.id}
-                    formId={parameterJSON.formId}
-                    formIdToLoad={parameterJSON.formIdToLoad}
-                    displayName={parameterJSON.displayName}
-                    value={parameterJSON.value}
-                    externalChannelName={parameterJSON.externalChannelName}
-                    selectionChanged={(action) => updateEditedJSON(action, parameterJSON.formId)}
+                    key={param.id}
+                    editorType={param.editorType}
+                    id={param.id}
+                    formId={param.formId}
+                    formIdToLoad={param.formIdToLoad}
+                    displayName={param.displayName}
+                    value={param.value}
+                    externalChannelName={param.externalChannelName}
+                    selectionChanged={(action) => updateEditedJSON(action, param.formId)}
                 />
             )}
         </StackLayout>
