@@ -1,13 +1,8 @@
-﻿import UPDATE from '../actions/formParams/update';
+﻿import UPDATE from "../actions/formParams/update";
+
 
 function updateParam(formId, id, value, manual) {
-    return {
-        type: UPDATE,
-        formId: formId,
-        id: id,
-        value: value,
-        manual: manual
-    };
+    return {type: UPDATE, formId, id, value, manual};
 }
 
 export default updateParam;
