@@ -114,7 +114,7 @@ module.exports = {
             },
             {
                 test: /\.m?(js|jsx)$/,
-                exclude: /(node_modules|bower_components)/,
+                exclude: /node_modules/,
                 use: {
                     loader: 'babel-loader',
                     options: {
@@ -125,6 +125,11 @@ module.exports = {
                         ]
                     }
                 }
+            },
+            {
+                test: /\.(ts|tsx)$/,
+                use: 'ts-loader',
+                exclude: /node_modules/,
             },
         ],
     },
